Move stuff out of api_types -> schema
