+++ testing/upload_validation.ts
@@ -13,20 +13,20 @@ import type {BooleanSchema} from '../schema';
 import type {CategorizationIndexDefinition} from '../schema';
 import type {CodaApiBearerTokenAuthentication} from '../types';
 import type {CodaInternalRichTextSchema} from '../schema';
-import type {CommentContentCategorization} from '../api_types';
+import type {CommentContentCategorization} from '../schema';
 import {ConnectionRequirement} from '../api_types';
-import {ContentCategorizationType} from '../api_types';
+import {ContentCategorizationType} from '../schema';
 import {CurrencyFormat} from '../schema';
 import type {CurrencySchema} from '../schema';
 import type {CustomAuthentication} from '../types';
 import type {CustomHeaderTokenAuthentication} from '../types';
 import type {CustomIndexDefinition} from '../schema';
 import type {DetailedIndexedProperty} from '../schema';
-import type {DocumentContentCategorization} from '../api_types';
+import type {DocumentContentCategorization} from '../schema';
 import type {DurationSchema} from '../schema';
 import {DurationUnit} from '../schema';
 import type {DynamicSyncTableDef} from '../api';
-import type {EmailContentCategorization} from '../api_types';
+import type {EmailContentCategorization} from '../schema';
 import {EmailDisplayType} from '../schema';
 import type {EmailSchema} from '../schema';
 import {FeatureSet} from '../types';
@@ -45,7 +45,7 @@ import {JSONPath} from 'jsonpath-plus';
 import {LifecycleBehavior} from '../schema';
 import {LinkDisplayType} from '../schema';
 import type {LinkSchema} from '../schema';
-import type {MessagingContentCategorization} from '../api_types';
+import type {MessagingContentCategorization} from '../schema';
 import type {MultiHeaderTokenAuthentication} from '../types';
 import type {MultiQueryParamTokenAuthentication} from '../types';
 import type {Network} from '../api_types';
